+++ rye/src/cli/rye.rs
@@ -21,15 +21,15 @@ use crate::bootstrap::{
 use crate::cli::toolchain::register_toolchain;
 use crate::config::Config;
 use crate::platform::{get_app_dir, symlinks_supported};
-use crate::sources::{get_download_url, PythonVersionRequest};
-use crate::utils::{check_checksum, toml, tui_theme, CommandOutput, QuietExit};
+use crate::sources::py::{get_download_url, PythonVersionRequest};
+use crate::utils::{check_checksum, toml, tui_theme, CommandOutput, IoPathContext, QuietExit};
 
 #[cfg(windows)]
 const DEFAULT_HOME: &str = "%USERPROFILE%\\.rye";
 #[cfg(unix)]
 const DEFAULT_HOME: &str = "$HOME/.rye";
 
-const GITHUB_REPO: &str = "https://github.com/mitsuhiko/rye";
+const GITHUB_REPO: &str = "https://github.com/astral-sh/rye";
 const UNIX_ENV_FILE: &str = r#"
 # rye shell setup
 {%- if custom_home %}
@@ -197,7 +197,7 @@ fn update(args: UpdateCommand) -> Result<(), Error> {
         let tmp = tempdir()?;
         cmd.arg("install")
             .arg("--git")
-            .arg("https://github.com/mitsuhiko/rye")
+            .arg("https://github.com/astral-sh/rye")
             .arg("--root")
             .env(
                 "PATH",
@@ -313,7 +313,7 @@ fn install(args: InstallCommand) -> Result<(), Error> {
 
 fn remove_dir_all_if_exists(path: &Path) -> Result<(), Error> {
     if path.is_dir() {
-        fs::remove_dir_all(path)?;
+        fs::remove_dir_all(path).path_context(path, "failed to remove directory")?;
     }
     Ok(())
 }
@@ -344,6 +344,7 @@ fn uninstall(args: UninstallCommand) -> Result<(), Error> {
         remove_dir_all_if_exists(&app_dir.join("self"))?;
         remove_dir_all_if_exists(&app_dir.join("py"))?;
         remove_dir_all_if_exists(&app_dir.join("pip-tools"))?;
+        remove_dir_all_if_exists(&app_dir.join("uv"))?;
         remove_dir_all_if_exists(&app_dir.join("tools"))?;
 
         // special deleting logic if we are placed in the app dir and the shim deletion
@@ -388,9 +389,15 @@ fn uninstall(args: UninstallCommand) -> Result<(), Error> {
 }
 
 #[cfg(unix)]
-fn is_fish() -> bool {
-    use whattheshell::Shell;
-    Shell::infer().map_or(false, |x| matches!(x, Shell::Fish))
+fn has_fish() -> bool {
+    use which::which;
+    which("fish").is_ok()
+}
+
+#[cfg(unix)]
+fn has_zsh() -> bool {
+    use which::which;
+    which("zsh").is_ok()
 }
 
 fn perform_install(
@@ -542,9 +549,9 @@ fn perform_install(
     // place executable in rye home folder
     fs::create_dir_all(&shims).ok();
     if target.is_file() {
-        fs::remove_file(&target)?;
+        fs::remove_file(&target).path_context(&target, "failed to delete old executable")?;
     }
-    fs::copy(exe, &target)?;
+    fs::copy(&exe, &target).path_context(&exe, "failed to copy executable")?;
     echo!("Installed binary to {}", style(target.display()).cyan());
 
     // write an env file we can source later.  Prefer $HOME/.rye over
@@ -554,10 +561,9 @@ fn perform_install(
         .unwrap_or((false, Cow::Borrowed(DEFAULT_HOME)));
 
     if cfg!(unix) {
-        fs::write(
-            app_dir.join("env"),
-            render!(UNIX_ENV_FILE, custom_home, rye_home),
-        )?;
+        let env_path = app_dir.join("env");
+        fs::write(&env_path, render!(UNIX_ENV_FILE, custom_home, rye_home))
+            .path_context(&env_path, "failed to write env file")?;
     }
 
     // Register a toolchain if provided.
@@ -666,7 +672,13 @@ fn add_rye_to_path(mode: &InstallMode, shims: &Path, ask: bool) -> Result<(), Er
             echo!();
             echo!("    source \"{}/env\"", rye_home.display());
             echo!();
-            if is_fish() {
+            if has_zsh() {
+                echo!("To make it work with zsh, you might need to add this to your .zprofile:");
+                echo!();
+                echo!("    source \"{}/env\"", rye_home.display());
+                echo!();
+            }
+            if has_fish() {
                 echo!("To make it work with fish, run this once instead:");
                 echo!();
                 echo!(