+++ rye/tests/common/mod.rs
@@ -25,7 +25,7 @@ pub const INSTA_FILTERS: &[(&str, &str)] = &[
     // windows temp folders
     (r"\b[A-Z]:\\.*\\Local\\Temp\\\S+", "[TEMP_FILE]"),
     (r" in (\d+\.)?\d+(ms|s)\b", " in [EXECUTION_TIME]"),
-    (r"\\([\w\d.])", "/$1"),
+    (r"\\\\?([\w\d.])", "/$1"),
     (r"rye.exe", "rye"),
 ];
 
@@ -50,14 +50,14 @@ fn bootstrap_test_rye() -> PathBuf {
 use-uv = true
 
 [default]
-toolchain = "cpython@3.12.1"
+toolchain = "cpython@3.12.2"
 "#,
         )
         .unwrap();
     }
 
     // fetch the most important interpreters
-    for version in ["cpython@3.8.17", "cpython@3.11.7", "cpython@3.12.1"] {
+    for version in ["cpython@3.8.17", "cpython@3.11.8", "cpython@3.12.2"] {
         if home.join("py").join(version).is_dir() {
             continue;
         }
@@ -152,6 +152,12 @@ impl Space {
     }
 
     #[allow(unused)]
+    pub fn read_toml<P: AsRef<Path>>(&self, path: P) -> toml_edit::Document {
+        let p = self.project_path().join(path.as_ref());
+        std::fs::read_to_string(p).unwrap().parse().unwrap()
+    }
+
+    #[allow(unused)]
     pub fn write<P: AsRef<Path>, B: AsRef<[u8]>>(&self, path: P, contents: B) {
         let p = self.project_path().join(path.as_ref());
         fs::create_dir_all(p.parent().unwrap()).ok();
@@ -184,6 +190,11 @@ impl Space {
 
     pub fn project_path(&self) -> &Path {
         &self.project_dir
+    }
+
+    #[allow(unused)]
+    pub fn venv_path(&self) -> PathBuf {
+        self.project_dir.join(".venv")
     }
 
     #[allow(unused)]